% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignMatch.R
\name{assignMatch}
\alias{assignMatch}
\title{Convert LoR-API call to a wide tibble}
\usage{
assignMatch(APIcall)
}
\arguments{
\item{APIcall}{\itemize{
\item a REST API response
}}
}
\value{
a dataframe with the metadata of a match
}
\description{
Convert LoR-API call to a wide tibble
}
