% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute_db_query.R
\name{execute_db_query}
\alias{execute_db_query}
\title{Execute a query in MySQL}
\usage{
execute_db_query(
  query,
  print_text = FALSE,
  conn = NULL,
  bigint = "integer64",
  ...
)
}
\arguments{
\item{query}{The query to execute (either the path to a .sql file, or a character string)}

\item{print_text}{Print query text? default: false}

\item{conn}{connection to the db; if not provided, create one and close it after the query}

\item{bigint}{Class to use for numeric values (one of integer", "numeric", "character", "integer64")}

\item{...}{Additional parameters to use for parametric queries}
}
\description{
Execute a query in MySQL
}
