% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_account_by_puuid.R
\name{get_account_by_puuid}
\alias{get_account_by_puuid}
\title{Given the primary key, a puuid, return all the ACCOUNT-V1 info
Wraps the \href{"https://developer.riotgames.com/apis#account-v1/GET_getByRiotId"}{GET_getByRiotId} api method and the \href{"https://developer.riotgames.com/apis#account-v1/GET_getActiveShard"}{GET_getActiveShard} api method.}
\usage{
get_account_by_puuid(puuid, server = "americas")
}
\arguments{
\item{puuid}{a character, string for PUUID, a string of 42char like RGAPI-XXXXXXXX-XXXX-XXXX-XXXX-XXXXXXXXXXXX}

\item{server}{a character, must be one of americas,europe,sea or asia,apac}
}
\description{
Standard RATE LIMITS
\emph{20 requests every 1 seconds(s) / 100 requests every 2 minutes(s) - Developer Key}
\emph{500 requests every 10 seconds  / 30,000 requests every 10 minutes - Production Key}
}
\details{
Method RATE LIMITS
\emph{1000 requests every 1 minutes - Developer Key}
\emph{1000 requests every 1 minutes - Production Key}
}
\examples{
\dontrun{
puuid <- "kJKtE_3i_66edP3lUYSW3wOVxIl5sRKFhsF6IpNIX_RQxYmyBZxG94gNuR4dUe-ofBq_zy5Yll_gSw"
get_account_by_puuid(puuid = puuid)
}
}
