% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_archetype_from_deck_code.R
\name{get_archetype_from_deck_code}
\alias{get_archetype_from_deck_code}
\title{Get an archetype name from a deck code}
\usage{
get_archetype_from_deck_code(deck_code, engine = "r")
}
\arguments{
\item{deck_code}{A character string with the code of the deck}

\item{engine}{One of "r" or "python" (default: "r")}
}
\value{
A character string
}
\description{
Get an archetype name from a deck code
}
