% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_current_patch.R
\name{get_current_patch}
\alias{get_current_patch}
\title{Get info on the most recent patch released}
\usage{
get_current_patch(
  what = "patch",
  since_last_change = TRUE,
  collapse = since_last_change,
  sep = "|"
)
}
\arguments{
\item{what}{One of "patch", "patch_regex", "release_date" (default: "patch")}

\item{since_last_change}{If TRUE, returns all patches since the last one with changes; otherwise returns only the very last one (default: TRUE)}

\item{collapse}{If since_last_change = TRUE, should patches be collapsed into a single character string? (default: TRUE)}

\item{sep}{separator to use for collapsing when collapse = TRUE (default: "|")}
}
\value{
A character string (if since_last_change = FALSE or collapse = TRUE); otherwise character vector
}
\description{
Get info on the most recent patch released
}
