% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_archetype_from_deck_code.R
\name{get_monoregion}
\alias{get_monoregion}
\title{Extract region from mono-region champions}
\usage{
get_monoregion(champs, data_champs = NULL)
}
\arguments{
\item{champs}{A string with a list of champion codes}

\item{data_champs}{Champions data from LoR Data Dragon (if NULL, the default, it will be fetched within the function)}
}
\value{
A character string with the list of regions from monoregion champs
}
\description{
Extract region from mono-region champions
}
