% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_current_patch.R
\name{get_patch_release_date}
\alias{get_patch_release_date}
\title{Get info on the release date of a patch}
\usage{
get_patch_release_date(patch = NULL, on_live = TRUE)
}
\arguments{
\item{patch}{The patch to check, in the format "live_X_XX" (if NULL, uses the latest patch since changes)}

\item{on_live}{If TRUE, adds 1 day to return the date when the patch actually went live (default: TRUE)}
}
\value{
A date
}
\description{
Get info on the release date of a patch
}
