% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_db_query.R
\name{get_query_statement}
\alias{get_query_statement}
\title{Read query from a SQL file (or from a charcter string)}
\usage{
get_query_statement(query)
}
\arguments{
\item{query}{The query to read (either a file path, or a character string)}
}
\value{
A character string, ready to be read by R
}
\description{
Read query from a SQL file (or from a charcter string)
}
