% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_regions_data.R
\name{get_regions_data}
\alias{get_regions_data}
\title{Get all data about LoR regions}
\usage{
get_regions_data(select = NULL)
}
\arguments{
\item{select}{A character vector with the fields to keep (return all fields if NULL)}
}
\value{
A dataframe
}
\description{
Get all data about LoR regions
}
\examples{
get_regions_data()
}
