% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cards_data.R
\name{get_set_cards_data}
\alias{get_set_cards_data}
\title{Get all data about cards from a specific set}
\usage{
get_set_cards_data(set, select = NULL)
}
\arguments{
\item{set}{The number of the set}

\item{select}{A character vector with the fields to keep (return all fields if NULL)}
}
\value{
A dataframe
}
\description{
Get all data about cards from a specific set
}
\examples{
get_set_cards_data(set = 1)
}
