% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cards_data.R
\name{get_sets}
\alias{get_sets}
\title{Get all sets released}
\usage{
get_sets(trim = TRUE)
}
\arguments{
\item{trim}{boolean; shold the "Set" prefix be removed? (default: true)}
}
\value{
A vector with all released sets
}
\description{
Get all sets released
}
