% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_scripts_in_folder.R
\name{launch_scripts_in_folder}
\alias{launch_scripts_in_folder}
\title{Run all scripts inside a folder, sending a discord message if they fail}
\usage{
launch_scripts_in_folder(
  folder,
  log_time = TRUE,
  log_message = "End: ",
  send_error_to_discord = TRUE,
  discord_username,
  discord_message = "Error. "
)
}
\arguments{
\item{folder}{Path to the folder with the scripts to run}

\item{log_time}{boolean, should we keep track of the time needed to run the script? (default: TRUE)}

\item{log_message}{Message to send alongside the elapsed time in logs}

\item{send_error_to_discord}{boolean, should a discord message be sent if the scripts fail? (default: TRUE)}

\item{discord_username}{The username that will be shown as sender on Discord}

\item{discord_message}{The content of the message sent on Discord}
}
\value{
Nothing, just runs the scripts
}
\description{
Run all scripts inside a folder, sending a discord message if they fail
}
