% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lor_leaderboard_dupe.R
\name{lor_leaderboard_dupe}
\alias{lor_leaderboard_dupe}
\title{Return Leaderboard Duplicate}
\usage{
lor_leaderboard_dupe(server, api_key = Sys.getenv("LORAPI_KEY"))
}
\arguments{
\item{server}{a character, must be one of "americas","apac","europe","asia","sea"
Note; asia and sea are going to be removed after the 20th of January 2022}

\item{api_key}{a character, default is an environment variables called LORAPI which must be set-up.}
}
\value{
a character vector
}
\description{
Return the duplicated players' names
Wraps the \href{https://developer.riotgames.com/apis#lor-ranked-v1/GET_getLeaderboards}{leaderboard} api method.
The leaderboard is updated once an hour.
}
\details{
Standard RATE LIMITS
\emph{20 requests every 1 seconds(s) / 100 requests every 2 minutes(s) - Developer Key}
\emph{500 requests every 10 seconds  / 30,000 requests every 10 minutes - Production Key}

Method RATE LIMITS
\emph{X-Method-Rate-Limit": "30:10,500:600 - Developer Key}
\emph{X-Method-Rate-Limit": "500:10,30000:600 - Production Key}
}
\examples{
\dontrun{
lor_leaderboard_dupe("europe")
lor_leaderboard_dupe("americas")
lor_leaderboard_dupe("asia")
}
}
