% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lor_match_list.R
\name{lor_match_list}
\alias{lor_match_list}
\title{Get the Match History}
\usage{
lor_match_list(server, puuid, maxPause = 10, wait = T, format = "parsed", ...)
}
\arguments{
\item{server, }{a character, must be one of americas,europe or sea.}

\item{puuid}{a character, string for PUUID, a string of 42char like RGAPI-XXXXXXXX-XXXX-XXXX-XXXX-XXXXXXXXXXXX}

\item{maxPause}{a numeric, in case of a call with status 429, what's the max wait it can take? default is 10s.
With a developer key 120 is the suggested.}

\item{wait}{a logical, if TRUE (the default), if the pause is of less or equal to 10s it waits and repeat the call once}

\item{format}{a character, format of the output, must be:
parsed - tibble of n games rows
text   - as the original json from the API request}

\item{...}{additional paramter for RETRY function, at the moment are timeout, times, pause_base, pause_cap, pause_min,}
}
\value{
a tibble, contains three variables:
\itemize{
\item match_id a character,
\item puuid  a character, string for PUUID, a string of 42char like RGAPI-XXXXXXXX-XXXX-XXXX-XXXX-XXXXXXXXXXXX
\item server a character, should be one of "americas", "asia", "europe"
}
}
\description{
Given a server amd primary key, a PUUID, return the Match History of a player (max 20)
Wraps the \href{"https://developer.riotgames.com/apis#lor-match-v1/GET_getMatchIdsByPUUID"}{GET_getMatchIdsByPUUID} api method.
}
\details{
Standard RATE LIMITS
\emph{20 requests every 1 seconds(s) / 100 requests every 2 minutes(s) - Developer Key}
\emph{500 requests every 10 seconds  / 30,000 requests every 10 minutes - Production Key}

Method RATE LIMITS
\emph{200:3600 - 200 requests every 1 hours   - Developer Key}
\emph{30:10    - 30 requests every 10 seconds - Production Key} (not fixed ratio for all Production Keys)
}
\examples{
\dontrun{
server <- "apac"
puuid <- "WE83ZhtKkg_nUWQFSVBlCadvrDAuDZag1q8c_VXRvIs-ttdCP7XscSCsK59ddn2WGVRYq5olFpIW5w"
lor_match_list(server=server,puuid=puuid)
lor_match_list(server=server,puuid=puuid,format="text")
}
}
