% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lor_match_metadata.R
\name{lor_match_metadata}
\alias{lor_match_metadata}
\title{Get the Match Metadata}
\usage{
lor_match_metadata(
  server,
  match_id,
  format = "parsed",
  maxPause = 10,
  wait = T,
  verbose = T,
  ...
)
}
\arguments{
\item{server, }{a character, must be one of americas,europe,sea or asia,apac}

\item{match_id}{a character, string for match_id, a string of 36char}

\item{format}{a character, format of the output, must be:
parsed: tibble of n row for n match
long: tibble of p row for each participant in a match
text: as the original json from the API request}

\item{maxPause}{a numeric, in case of a call with status 429, what's the max wait it can take? default is 10s.
With a developer key 120 is the suggested.}

\item{wait}{a logical, if TRUE (the default), if the pause is of less or equal to 10s it waits and repeat the call once}

\item{verbose}{should be function be verbose and print messages}

\item{...}{additional paramter for RETRY function, at the moment are timeout, times, pause_base, pause_cap, pause_min,}
}
\value{
a tibble
}
\description{
Given a server amd a match_id, return the Match Metadata of a match
Wraps the \href{"https://developer.riotgames.com/apis#lor-match-v1/GET_getMatch"}{GET_getMatch} api method.
}
\details{
Standard RATE LIMITS
\emph{20 requests every 1 seconds(s) / 100 requests every 2 minutes(s) - Developer Key}
\emph{500 requests every 10 seconds  / 30,000 requests every 10 minutes - Production Key}

Method RATE LIMITS
\emph{100:3600 - 200 requests every 1 hours   - Developer Key}
\emph{30:10    - 30 requests every 10 seconds - Production Key} (not fixed ratio for all Production Keys)
}
\examples{
\dontrun{
server <-  "europe"
match_id <- "44a130ae-12f4-45f8-8a24-0b319265d616"
lor_match_metadata(server=server,match_id=match_id)
lor_match_metadata(server=server,match_id=match_id,format="text")
lor_match_metadata(server=server,match_id=match_id,format="long")
}
}
