% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newPuuids.R
\name{newPuuids}
\alias{newPuuids}
\title{Search for new PUUID}
\usage{
newPuuids(LoR.Account.df, LoR.Match.df)
}
\arguments{
\item{LoR.Account.df}{a tidy data frame, containing a column called \strong{puuid}}

\item{LoR.Match.df}{a tidy data frame, containing column(s) containng \strong{puuid} and one called \strong{game_mode}}
}
\value{
a vector of character, containing the puuid in the matches dataframe not already present in the account data frame
}
\description{
Check if there are any new PUUID from several game modes (those not against bot) and return a vector of PUUID that are not included
"Constructed","SeasonalTournamentLobby","Bo3ChallengeLobby","StandardGauntletLobby","LastCallQualifierGauntletLobby"
}
\examples{

Example.Account.df = data.frame(
	gameName = c("Pippo","Pluto"),
	puuid = c("defg","wxyz")
)

Example.Match.df = data.frame(
	game_mode = c("Constructed","Constructed","AI"),
	puuid_1 = c("abcd","defg","mnop"),
	puuid_2 = c("defg","hijk","bot1"),
	puuid_3 = c("wxyz",NA,NA),
	puuid_4 = c(NA,NA,NA)
)
newPuuids(LoR.Account.df = Example.Account.df, LoR.Match.df = Example.Match.df)
}
