% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_discord_message.R
\name{send_discord_message}
\alias{send_discord_message}
\title{Send message to discord}
\usage{
send_discord_message(username, message)
}
\arguments{
\item{username}{The username that will be shown as sender on Discord}

\item{message}{The content of the message}
}
\description{
Send message to discord
}
