% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setAPIkey.R
\name{setAPIkey}
\alias{setAPIkey}
\title{Setup the api}
\usage{
setAPIkey(api_key)
}
\arguments{
\item{api_key}{a character, it should have length 42}
}
\value{
a global variable # invisible
}
\description{
Create a LORAPI_KEY variable that contains the lor-API key.
LORAPI_KEY is used as default value for functions which calls the API
}
\examples{
setAPIkey("RGAPI-01234567-abcd-edfg-hijk-01234567890A")
}
