% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_call.R
\name{api_call}
\alias{api_call}
\title{LoR-API Handler}
\usage{
api_call(server, path, ..., api_key = Sys.getenv("LORAPI_KEY"))
}
\arguments{
\item{server, }{a character, must be one of "americas", "asia", "europe", "sea"}

\item{path}{a character, string containing the path of the methods of choice. Please refer to the \href{https://developer.riotgames.com/apis}{developer portal}}

\item{...}{all arguments accepted by httr::RETRY like: httr::timeout(timeout), times, quiet,}

\item{api_key}{a environment variables.}
}
\value{
a response
}
\description{
This function helps you handling the REST API for the LoR API methods
}
\details{
#' Standard RATE LIMITS
\emph{20 requests every 1 seconds(s) / 100 requests every 2 minutes(s) - Developer Key}
\emph{500 requests every 10 seconds  / 30,000 requests every 10 minutes - Production Key}
}
\examples{
api_call("europe","/lor/ranked/v1/leaderboards/") # not run
api_call("europe","/lor/ranked/v1/leaderboards/",timeout(3),times=3,quiet=FALSE) # not run
}
