% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignPlayer.R
\name{assignPlayer}
\alias{assignPlayer}
\title{Convert LoR-API call player field to a wide tibble}
\usage{
assignPlayer(playerMetadata)
}
\arguments{
\item{playerMetadata}{\itemize{
\item player info from a LoR REST API response match metadata
}}
}
\value{

}
\description{
Convert LoR-API call player field to a wide tibble
}
