% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cards_data.R
\name{get_cards_data}
\alias{get_cards_data}
\title{Get all data about cards}
\usage{
get_cards_data(select = NULL)
}
\arguments{
\item{select}{A character vector with the fields to keep (return all fields if NULL)}
}
\value{
A dataframe
}
\description{
Get all data about cards
}
\examples{
get_cards_data()
}
