% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_db_query.R
\name{get_db_query}
\alias{get_db_query}
\title{Get data from a query in MySQL}
\usage{
get_db_query(
  query,
  limit = -1,
  print_text = FALSE,
  print_df = FALSE,
  conn = NULL
)
}
\arguments{
\item{query}{The query to execute}

\item{limit}{Maximum number of rows to retrive default: unbounded}

\item{print_text}{Print query text? default: false}

\item{print_df}{Print first rows of df? default: false}

\item{conn}{connection to the db; if not provided, create one and close it after the query}
}
\value{
A dataframe
}
\description{
Get data from a query in MySQL
}
