% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_path_server.R
\name{get_path_server}
\alias{get_path_server}
\title{Return the base path for one of Riot's shards}
\usage{
get_path_server(server)
}
\arguments{
\item{server}{a character, must be one of "americas","apac","europe","asia","sea"
Note; asia and sea are going to be removed after the 20th of January 2022}
}
\value{
a character, string containing the url
}
\description{
Return the base path for one of Riot's shards
}
\examples{
get_path_server("europe")
get_path_server("apac")
}
