% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_path_server.R
\name{get_path_server}
\alias{get_path_server}
\title{Return the base path for one of Riot's shards}
\usage{
get_path_server(server)
}
\arguments{
\item{server}{character - the value should be one among: americas,asia,europe,sea}
}
\value{
character string containing the url
}
\description{
Return the base path for one of Riot's shards
}
\examples{
get_path_server("europe")
get_path_server("americas")
}
